タイピングのカスタムフックでタイピングに関わるstateも管理するようにし、呼び出しを最上位のコンポーネントからとした
