+++ src/components/typing/ResetButton.tsx
@@ -1,34 +1,11 @@
-import { TypingStatus } from "@/types/typingStatus";
-
 type ResetButtonProps = {
-  setCursorLine: (cursorLine: number) => void;
-  setTypedTextLines: (typedTexts: string[]) => void;
-  setCursorPositions: (cursorPositions: number[]) => void;
-  initialTypedTextLines: string[];
-  initialCursorPositions: number[];
-  setTypingStatus: (typingStatus: TypingStatus) => void;
+  resetTyping: () => void;
 };
 
-export default function ResetButton({
-  setCursorLine,
-  setTypedTextLines,
-  setCursorPositions,
-  initialTypedTextLines,
-  initialCursorPositions,
-  setTypingStatus,
-}: ResetButtonProps) {
+export default function ResetButton({ resetTyping }: ResetButtonProps) {
   return (
     <>
-      <button
-        onClick={() => {
-          setCursorLine(0);
-          setTypedTextLines(initialTypedTextLines);
-          setCursorPositions(initialCursorPositions);
-          setTypingStatus("idling");
-        }}
-      >
-        リセット
-      </button>
+      <button onClick={resetTyping}>リセット</button>
     </>
   );
 }
