+++ src/pages/typings/index.tsx
@@ -5,25 +5,21 @@ import ResetButton from "@/components/typing/ResetButton";
 import Progress from "@/components/typing/Progress";
 import Result from "@/components/typing/Result";
 import Start from "@/components/typing/Start";
-import { TypingStatus } from "@/types/typingStatus";
+import { useTyping } from "@/hooks/useTyping";
 
 const Typing: NextPage<{}> = () => {
   const targetText = "def fizz_buzz(num)\n  if num % 15 == 0\nend";
   const targetTextLines = targetText.split(/(?<=\n)/);
-
-  // 行頭のスペースは入力済みとする
-  const initialCursorPositions = targetTextLines.map((line) =>
-    line.indexOf(line.trimStart())
-  );
-  const [cursorPositions, setCursorPositions] = useState(
-    initialCursorPositions
-  );
-  const initialTypedTextLines = targetTextLines.map((_, index) =>
-    " ".repeat(initialCursorPositions[index])
-  );
-  const [typedTextLines, setTypedTextLines] = useState(initialTypedTextLines);
-  const [typingStatus, setTypingStatus] = useState<TypingStatus>("idling");
-  const [cursorLine, setCursorLine] = useState(0);
+  const {
+    typedTextLines,
+    cursorPositions,
+    cursorLine,
+    typingStatus,
+    startTyping,
+    resetTyping,
+  } = useTyping({
+    targetTextLines,
+  });
 
   const typedText = typedTextLines.slice(0, cursorLine + 1).join("");
   const correctTypedTextCount = typedText
@@ -41,7 +37,7 @@ const Typing: NextPage<{}> = () => {
     <div>
       <h1>タイピング</h1>
       {typingStatus === "idling" ? (
-        <Start setTypingStatus={setTypingStatus} />
+        <Start startTyping={startTyping} />
       ) : typingStatus === "completed" ? (
         <Result
           targetTextLines={targetTextLines}
@@ -53,21 +49,10 @@ const Typing: NextPage<{}> = () => {
           <TypingArea
             targetTextLines={targetTextLines}
             typedTextLines={typedTextLines}
-            setTypedTextLines={setTypedTextLines}
             cursorPositions={cursorPositions}
-            setCursorPositions={setCursorPositions}
             cursorLine={cursorLine}
-            setCursorLine={setCursorLine}
-            setTypingStatus={setTypingStatus}
           />
-          <ResetButton
-            setCursorLine={setCursorLine}
-            setTypedTextLines={setTypedTextLines}
-            setCursorPositions={setCursorPositions}
-            initialTypedTextLines={initialTypedTextLines}
-            initialCursorPositions={initialCursorPositions}
-            setTypingStatus={setTypingStatus}
-          />
+          <ResetButton resetTyping={resetTyping} />
           <Progress
             typedTextCount={typedTextCount}
             targetTextCount={targetTextCount}
